Updated German translation texts
Added some translation texts that were missing in the German version of the website.

Updated some translation texts to be in sync the latest English version.

Updated some texts to be a bit more accurate or to fix some minor typos  and style issues.

Reordered some of the texts to match the order in which they appear on the website.